requirements.txt
# Base libraries
absl-py~=0.13.0
h5py~=3.1.0
ipykernel~=5.5.5
ipython~=7.24.1
ipython-genutils~=0.2.0
ipywidgets~=7.6.3
joblib~=1.0.1
numpy~=1.19.5
pillow~=8.2.0
psutil~=5.8.0
pytest~=6.2.4
python-snappy~=0.6.0
scipy~=1.6.3
tabulate~=0.8.9

# Base dependencies
argon2-cffi~=20.1.0
async-generator~=1.10
attrs~=21.2.0
backcall~=0.2.0
bleach~=3.3.0
cached-property~=1.5.2
certifi~=2021.5.30
cffi~=1.14.5
chardet~=4.0.0
click~=8.0.1
click-plugins~=1.1.1
cligj~=0.7.2
cloudpickle~=1.6.0
cycler~=0.10.0
decorator~=5.0.9
defusedxml~=0.7.1
dill~=0.3.4
entrypoints~=0.3
fiona~=1.8.20
flatbuffers~=2.0
idna~=2.10
importlib-metadata~=4.5.0
iniconfig~=1.1.1
itsdangerous~=2.0.1
jedi~=0.18.0
jinja2~=3.0.1
jsonschema~=3.2.0
kiwisolver~=1.3.1
markupsafe~=2.0.1
matplotlib-inline~=0.1.2
mistune~=0.8.4
munch~=2.5.0
nbclient~=0.5.3
nbconvert~=6.0.7
nbformat~=5.1.3
nest-asyncio~=1.5.1
notebook~=6.4.0
opt-einsum~=3.3.0
packaging~=20.9
pandocfilters~=1.4.3
parso~=0.8.2
pexpect~=4.8.0
pickleshare~=0.7.5
pluggy~=0.13.1
prometheus-client~=0.11.0
prompt-toolkit~=3.0.18
ptyprocess~=0.7.0
py~=1.10.0
py4j~=0.10.9
pycparser~=2.20
pyglet~=1.5.15
pygments~=2.9.0
pyparsing~=2.4.7
pyproj~=3.1.0
pyrsistent~=0.17.3
python-dateutil~=2.8.1
pyyaml~=5.4.1
qtconsole~=5.1.0
qtpy~=1.9.0
send2trash~=1.5.0
terminado~=0.10.1
testpath~=0.5.0
threadpoolctl~=2.1.0
toml~=0.10.2
toolz~=0.11.1
tornado~=6.1
traitlets~=5.0.5
typing-extensions~=3.10.0.0
urllib3~=1.26.5
wcwidth~=0.2.5
webencodings~=0.5.1
widgetsnbextension~=3.5.1
zipp~=3.4.1

# ODySSEUS libraries
bokeh~=2.3.2
flask~=1.1.2
geopandas~=0.8.2
haversine~=2.3.1
keras~=2.4.3
matplotlib~=3.3.4
mlrose~=1.3.0
pandas~=1.1.5
pathlib~=1.0.1
pickle5~=0.0.11
pyspark~=3.0.2
pytz~=2019.3
requests~=2.25.1
rtree~=0.9.7
scikit-learn~=0.23.2
seaborn~=0.9.1
shapely~=1.7.1
simpy~=4.0.1
six~=1.15.0
sklearn~=0.0
sodapy~=2.1.0
tqdm~=4.54.1
werkzeug~=2.0.1
xmltodict~=0.12.0

# Request-reply API libraries
pyzmq~=22.0.3

# DeepMind libraries
chex~=0.0.7
dm-env~=1.4
dm-haiku~=0.0.4
dm-tree~=0.1.6
optax~=0.0.8
rlax~=0.0.3

# JAX libraries with GPU support
jax~=0.2.14
jaxlib~=0.1.67

# ALE and Gym libraries
atari-py~=0.2.9
gym~=0.18.3

# Jupyter libraries
jupyter~=1.0.0
jupyter-client~=6.1.12
jupyter-console~=6.4.0
jupyter-core~=4.7.1
jupyterlab-pygments~=0.1.2
jupyterlab-widgets~=1.0.0
